pyproject.toml
[project]
name = "geomas"
version = "0.1.0"
description = "Tool for model training"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "PyPDF2",
    "timm",
    "bs4",
    "mlflow",
    "psutil",
    "pynvml",
    "boto3",
    "dotenv",
    "google-generativeai",
    "tenacity",
    # OCR dependencies
    "fpdf2>=2.8.4",
    "pillow>=10.2,<11",
    "python-pptx>=1.0.2",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
    "tqdm>=4.67.1,<5"
]

[tool.setuptools]
packages = ["geomas"]

[project.optional-dependencies]  # pip install -e ".[dev]"
dev = [
    "ruff",
    "pytest",
]
ocr = [
    # Advanced OCR adapters (optional due to heavy dependencies)
    "marker-pdf>=1.8.4",
    "magic-pdf[lite]>=1.3.12",
    "olmocr>=0.3.3",
    "torch>=2.5.1",
    "transformers>=4.53.0,<4.54.0",
    "huggingface-hub>=0.24.0",
    "fuzzysearch>=0.7.3",
    "playwright>=1.54.0",
    "ultralytics>=8.3.0",
    "doclayout-yolo>=0.0.4",
    "rapid-table>=2.0.3",
    "pypdfium2>=4.27.0",
    "accelerate>=0.34.2",
    "compressed-tensors>=0.5.1",
    "pymupdf>=1.24.10",
    "numpy>=1.26,<2",
]

[tool.ruff.lint]
ignore = ["E501", "F401"]  # Ignore long lines and unused imports

[project.scripts]
geomas = "geomas.cli:app"
